use new disconnect syntax
